// techstore-mobile/src/api/generated/types.gen.ts
export type MarcaResponse = {
    id: number;
    nombre: string;
};

export type ModeloRef = {
    id: number;
    nombre: string;
};

export type ProductoPublicResponse = {
    id: number;
    nombre: string;
    descripcion?: string | null;
    moneda: 'PYG' | 'USD';
    precio: number;
    iva: number;
    disponible: boolean;
    marca?: MarcaResponse | null;
    modelo?: ModeloRef | null;
    imagenUrl?: string | null;
};

export type ProductoPage = {
    content: Array<ProductoPublicResponse>;
    page: number;
    size: number;
    totalElements: number;
    totalPages: number;
};

export type FavoritoResponse = {
    producto: ProductoPublicResponse;
    creadoEn: string;
};